+++ package.json
@@ -43,6 +43,7 @@
     "eslint-plugin-regexp": "^1.5.1",
     "eslint-plugin-simple-import-sort": "^7.0.0",
     "eslint-plugin-unicorn": "^39.0.0",
+    "eslint-plugin-unused-imports": "^2.0.0",
     "husky": "^7.0.4",
     "jest": "^27.0.0",
     "lint-staged": "^12.1.3",
