--- src/generateRolls/modifiers/index.ts
@@ -1,5 +1,0 @@
-export * from './modify-roll-totals'
-export * from './modify-total'
-export * from './replace'
-export * from './reroll'
-export * from './unique'
