--- src/generateRolls/modifiers/minus/index.ts
@@ -1,1 +1,0 @@
-export * from './parse-minus'
