--- src/generateRolls/modifiers/plus/index.ts
@@ -1,1 +1,0 @@
-export * from './parse-plus'
