--- src/generateRolls/modifiers/replace/index.ts
@@ -1,1 +1,0 @@
-export * from './parse-replace'
