--- src/generateRolls/modifiers/reroll/index.ts
@@ -1,1 +1,0 @@
-export * from './parse-reroll'
