--- src/generateRolls/modifiers/unique/index.ts
@@ -1,1 +1,0 @@
-export * from './parse-unique'
