+++ src/generateRolls/parsers/modify-total.ts
@@ -1,7 +1,7 @@
 import { RollParameters } from 'types'
 
-import { parseMinusFactory } from './minus'
-import { parsePlusFactory } from './plus'
+import { parseMinusFactory } from './parse-minus'
+import { parsePlusFactory } from './parse-plus'
 
 export function modifyTotal(total: number, { plus, minus }: Pick<RollParameters, 'plus' | 'minus'>) {
   const modifiers = [parsePlusFactory(plus), parseMinusFactory(minus)]
