+++ src/generateRolls/parsers/parse-cap.ts
